+++ NEWS.md
@@ -1,5 +1,13 @@
 # MSnbase 2.5
 
+## Changes in version 2.5.13
+
+- Reduce unit testing time (see
+  [#334](https://github.com/lgatto/MSnbase/issues/334))
+  <2018-04-13 Fri>
+- Fix bug in write.exprs when only one feature data is passed
+  <2018-04-17 Tue>
+
 ## Changes in version 2.5.12
 
 - Ensure tic and bpi with initial = TRUE calculate the tic and bpi from the